+++ app/controllers/questions_controller.rb
@@ -32,9 +32,9 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    if current_user.author?(question)
+    if user_signed_in? && current_user.author?(question)
       question.destroy
-      flash[:notice] = 'Answer delete successfully'
+      flash[:notice] = 'Deleted successfully'
     else
       flash[:alert] = "You can not delete"
     end
